Handle supportsMediaUpload.
If there is a method that supportsMediaUpload, then we expand that
method into up to 3 separate paths (original, simple, and resumable).

The simple and resumable methods share most attributes with the original
method, while adding a few implied parameters (from the documentation).

Since the upload paths are relative to / and not the basePath, we must
also set the basePath to "/" and prepend the original basePath to each
non upload path.